requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.1; python_full_version >= '3.7.0'
django==4.2.6; python_version >= '3.8'
django-cors-headers==4.3.0; python_version >= '3.8'
django-dotenv==1.4.2
djangorestframework==3.14.0; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.3; sys_platform == 'win32'
urllib3==2.0.7; python_version >= '3.7'
